Новый файл tests.py. Тест добавления данных, сравнение количества объектов. Сравнение данных с новым JSON объектом. Тест требует отладки. Ошибка в пакете loader.py
