+++ Pereval/mount/urls.py
@@ -1,8 +1,9 @@
-from django.urls import path
-
-from .views import PerevalViewset
-
-urlpatterns = [
-    path('', PerevalViewset.as_view(), name='create'),
-    path('<int:pk>/', PerevalViewset.as_view(), name='partial_update'),
-]
+# from django.urls import path
+#
+# from .views import PerevalViewset
+#
+#
+# urlpatterns = [
+#     path('', PerevalViewset.as_view(), name='create'),
+#     path('<int:pk>/', PerevalViewset.as_view(), name='partial_update'),
+# ]
